Backed out 3 changesets (bug 1644706, bug 1644437, bug 1644368) for python related bustage CLOSED TREE
Backed out changeset 2a6645398c5a (bug 1644706)
Backed out changeset e695f5748e04 (bug 1644437)
Backed out changeset d615591507b9 (bug 1644368)

+++ python/mozperftest/mozperftest/test/browsertime/runner.py
@@ -10,13 +10,13 @@ import re
 import shutil
 from pathlib import Path
 
+from mozperftest.scriptinfo import ScriptInfo
 from mozperftest.utils import install_package
 from mozperftest.test.noderunner import NodeRunner
 from mozperftest.test.browsertime.setup import (
     system_prerequisites,
     append_system_env,
 )
-from mozperftest.test.browsertime.script import ScriptInfo
 
 
 BROWSERTIME_SRC_ROOT = Path(__file__).parent
@@ -89,7 +89,7 @@ class BrowsertimeRunner(NodeRunner):
         self._mach_context = mach_cmd._mach_context
         self.virtualenv_manager = mach_cmd.virtualenv_manager
         self._created_dirs = []
-        self._test_script = None
+        self._test_info = {}
         self._setup_helper = None
         self.get_binary_path = mach_cmd.get_binary_path
 
@@ -128,13 +128,8 @@ class BrowsertimeRunner(NodeRunner):
         super(BrowsertimeRunner, self).setup()
         install_url = self.get_arg("install-url")
 
-        tests = self.get_arg("tests", [])
-        if len(tests) != 1:
-            # we don't support auto-discovery (no test passed) or multiple
-            # tests here yet.
-            raise NotImplementedError()
-
-        self._test_script = ScriptInfo(tests[0])
+        if self.get_arg("tests"):
+            self._test_info = ScriptInfo(self.get_arg("tests")[0])
 
         # installing Python deps on the fly
         for dep in ("Pillow==%s" % PILLOW_VERSION, "pyssim==%s" % PYSSIM_VERSION):
@@ -142,7 +137,7 @@ class BrowsertimeRunner(NodeRunner):
 
         # check if the browsertime package has been deployed correctly
         # for this we just check for the browsertime directory presence
-        if self.browsertime_js.exists() and not self.get_arg("clobber"):
+        if self.browsertime_js.exists():
             return
 
         if install_url is None:
@@ -154,8 +149,8 @@ class BrowsertimeRunner(NodeRunner):
         for file in ("package.json", "package-lock.json"):
             src = BROWSERTIME_SRC_ROOT / file
             target = self.state_path / file
-            # Overwrite the existing files
-            shutil.copyfile(str(src), str(target))
+            if not target.exists():
+                shutil.copyfile(str(src), str(target))
 
         package_json_path = self.state_path / "package.json"
 
@@ -308,6 +303,7 @@ class BrowsertimeRunner(NodeRunner):
 
     def _one_cycle(self, metadata, result_dir):
         profile = self.get_arg("profile-directory")
+        test_script = self.get_arg("tests")[0]
 
         args = [
             "--resultDir",
@@ -316,7 +312,7 @@ class BrowsertimeRunner(NodeRunner):
             profile,
             "--iterations",
             str(self.get_arg("iterations")),
-            self._test_script["filename"],
+            test_script,
         ]
 
         if self.get_arg("verbose"):
@@ -330,12 +326,6 @@ class BrowsertimeRunner(NodeRunner):
                     continue
                 option = option.split("=")
                 if len(option) != 2:
-                    self.warning(
-                        f"Skipping browsertime option {option} as it "
-                        "is missing a name/value pairing. We expect options "
-                        "to be formatted as: --browsertime-extra-options "
-                        "'browserRestartTries=1,timeouts.browserStart=10'"
-                    )
                     continue
                 name, value = option
                 args += ["--" + name, value]
@@ -351,7 +341,10 @@ class BrowsertimeRunner(NodeRunner):
             raise NodeException(exit_code)
 
         metadata.add_result(
-            {"results": str(result_dir), "name": self._test_script["name"]}
+            {
+                "results": str(result_dir),
+                "name": self._test_info.get("name", ("browsertime",))[0],
+            }
         )
 
         return metadata
